Add `carrick run -d` + ps/stop/kill/rm (daemonless lifecycle)
Docker-style container lifecycle, with no central daemon — each detached
container is its own process tree under a per-container NsSupervisor, and the
lifecycle subcommands are pure reads/signals over the on-disk registry (podman
model; see memory carrick-no-daemon).

`carrick run -d`: generate a 64-hex id, write a Created registry entry, then
fork+setsid into the background with stdio redirected to the container's log,
print the id, and return. The detached child exports CARRICK_CONTAINER_ID and
runs the engine; the runtime's NsSupervisor records the live init/supervisor
pids (status → Running) and marks the entry Exited (or removes it for --rm)
when the init exits. Conflicts with -t/-i.

`carrick ps [-a] [-q]`, `stop [-t] <c…>`, `kill [-s SIG] <c…>`, `rm [-f] <c…>`:
resolve by id / name / unambiguous id-prefix, signal the recorded host init
pid, and reconcile a Running entry whose init has died to Exited.

Supporting fixes found by testing the lifecycle end-to-end:
- NsSupervisor deadlock: the registration-pipe drain did a blocking read on a
  blocking fd and wedged the supervisor after the first event (so SIGTERM to a
  container's init never triggered teardown). Both pipe ends are now
  non-blocking; the drain terminates on EAGAIN.
- pid-1 signal protection (pid_namespaces(7) §5.4), now load-bearing because the
  supervisor faithfully tears the container down on init exit: a default-lethal
  signal sent to the ns-init by another ns member is dropped unless the init
  installed a handler (SIGKILL/SIGSTOP always act). The init publishes its
  handler mask to the shared region from rt_sigaction; the kill path consults it.
- pid_alive now treats a zombie (host state 'Z') as dead, so a stopped
  container's unreaped init isn't reported "running" until launchd reaps it.
- container ids that reach a path are allowlist-validated (CWE-22).

Verified end-to-end on the signed build: run -d → ps (running) → stop/kill (by
id and name) → exited, 0 leftover guest procs → rm; -d --rm auto-removes on
exit; rm -f force-removes a running container; foreground exit-code passthrough
unchanged. Full conformance suite green (0 gaps), runtime unit suite 250 tests.

+++ crates/carrick-runtime/src/namespace/supervisor.rs
@@ -56,7 +56,7 @@ pub fn run(init_host_pid: i32, reg_pipe_read: i32) -> SupervisorExit {
     // Watch the registration pipe for new members, and arm an exit watch on the
     // init itself (so its death wakes the loop just like any other member).
     let _ = kq.apply(&[
-        Kevent::read(reg_pipe_read, libc::EV_ADD as u16),
+        Kevent::read(reg_pipe_read, libc::EV_ADD),
         Kevent::proc_exit(init_host_pid),
     ]);
